requirements.txt
fastapi
uvicorn
python-dotenv
pydantic
requests
neo-mamba
prisma
asyncpg
spoon_ai
